--- sentry_sdk/integrations/pydantic_ai/patches/__init__.py
@@ -1,3 +1,0 @@
-from .agent_run import _patch_agent_run  # noqa: F401
-from .graph_nodes import _patch_graph_nodes  # noqa: F401
-from .tools import _patch_tool_execution  # noqa: F401
