--- sentry_sdk/integrations/pydantic_ai/spans/__init__.py
@@ -1,3 +1,0 @@
-from .ai_client import ai_client_span, update_ai_client_span  # noqa: F401
-from .execute_tool import execute_tool_span, update_execute_tool_span  # noqa: F401
-from .invoke_agent import invoke_agent_span, update_invoke_agent_span  # noqa: F401
